minor: don't run copy_database w/ username & pass unless we're on 1.3.x

+++ test/test_connection.py
@@ -30,6 +30,7 @@ from pymongo.errors import (AutoReconnect,
                             InvalidName,
                             InvalidURI,
                             OperationFailure)
+from test import version
 
 
 def get_connection(*args, **kwargs):
@@ -159,23 +160,25 @@ class TestConnection(unittest.TestCase):
         c.drop_database("pymongo_test1")
         c.drop_database("pymongo_test2")
 
-        c.pymongo_test.add_user("mike", "password")
+        if version.at_least(c, (1, 3, 3, 1)):
+            c.pymongo_test.add_user("mike", "password")
 
-        self.assertRaises(OperationFailure, c.copy_database,
-                          "pymongo_test", "pymongo_test1",
-                          username="foo", password="bar")
-        self.failIf("pymongo_test1" in c.database_names())
+            self.assertRaises(OperationFailure, c.copy_database,
+                              "pymongo_test", "pymongo_test1",
+                              username="foo", password="bar")
+            self.failIf("pymongo_test1" in c.database_names())
 
-        self.assertRaises(OperationFailure, c.copy_database,
-                          "pymongo_test", "pymongo_test1",
-                          username="mike", password="bar")
-        self.failIf("pymongo_test1" in c.database_names())
+            self.assertRaises(OperationFailure, c.copy_database,
+                              "pymongo_test", "pymongo_test1",
+                              username="mike", password="bar")
+            self.failIf("pymongo_test1" in c.database_names())
 
-        c.copy_database("pymongo_test", "pymongo_test1", username="mike", password="password")
-        self.assert_("pymongo_test1" in c.database_names())
-        self.assertEqual("bar", c.pymongo_test1.test.find_one()["foo"])
+            c.copy_database("pymongo_test", "pymongo_test1",
+                            username="mike", password="password")
+            self.assert_("pymongo_test1" in c.database_names())
+            self.assertEqual("bar", c.pymongo_test1.test.find_one()["foo"])
 
-        c.drop_database("pymongo_test1")
+            c.drop_database("pymongo_test1")
 
     def test_iteration(self):
         connection = Connection(self.host, self.port)
